fix "resize_on_focus" feature and expose it in settings
"resize_on_focus" feature is not working, i.e. when
setting it to false the resize still happens after
clicking by mouse an unfocused group.

Probably that was the reason this feature was not exposed
in the default plugin settings file, because it was not
working.

So this commit makes that setting to work as expected and
it's added to the plugin settings file.